[SL-162] docs: 그룹멤버 조인테이블 관련 조회 API를 ChannelMember 컨트롤러로 이동
