+++ src/main/java/com/sds/actlongs/controller/channelmember/ChannelMemberController.java
@@ -1,8 +1,14 @@
 package com.sds.actlongs.controller.channelmember;
 
+import java.util.List;
+
 import javax.validation.Valid;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,12 +20,43 @@ import lombok.RequiredArgsConstructor;
 
 import com.sds.actlongs.controller.channelmember.dto.MemberInviteRequest;
 import com.sds.actlongs.controller.channelmember.dto.MemberInviteResponse;
+import com.sds.actlongs.controller.channelmember.dto.MemberListResponse;
+import com.sds.actlongs.controller.channelmember.dto.MemberSearchResponse;
 
 @Api(tags = "그룹회원 API")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/group-members")
 public class ChannelMemberController {
+
+	@ApiOperation(value = "그룹원 목록 조회 API", notes = "ML001: 그룹원 목록 조회에 성공하였습니다.")
+	@GetMapping("/{groupId}")
+	public ResponseEntity<MemberListResponse> findMemberList(@PathVariable("groupId") final Long channelId) {
+		List<MemberListResponse.MemberResponse> memberList = List.of(
+			new MemberListResponse.MemberResponse("Sean"),
+			new MemberListResponse.MemberResponse("Ari"),
+			new MemberListResponse.MemberResponse("Jin"),
+			new MemberListResponse.MemberResponse("Null")
+		);
+		MemberListResponse listResponse = new MemberListResponse(memberList);
+		return ResponseEntity.ok(listResponse);
+	}
+
+	@ApiOperation(value = "회원 검색 API", notes = "MS001: 회원 검색에 성공하였습니다.")
+	@GetMapping("/{groupId}/with/{keyword}/")
+	public ResponseEntity<MemberSearchResponse> searchMember(
+		@PathVariable @NotBlank @Size(max = 20) final String keyword,
+		@PathVariable("groupId") final Long channelId
+	) {
+		List<MemberSearchResponse.SearchedMember> searchList = List.of(
+			new MemberSearchResponse.SearchedMember(101L, "Din"),
+			new MemberSearchResponse.SearchedMember(102L, "diedie"),
+			new MemberSearchResponse.SearchedMember(103L, "DIN_DEAN"),
+			new MemberSearchResponse.SearchedMember(104L, "dIabcd")
+		);
+		MemberSearchResponse listResponse = new MemberSearchResponse(searchList);
+		return ResponseEntity.ok(listResponse);
+	}
 
 	@ApiOperation(value = "그룹원 초대 API", notes = "IV001: 그룹원 초대에 성공하였습니다.")
 	@PostMapping
