--- src/main/java/com/sds/actlongs/controller/member/dto/MemberListRequest.java
@@ -1,19 +1,0 @@
-package com.sds.actlongs.controller.member.dto;
-
-import javax.validation.constraints.NotNull;
-
-import io.swagger.annotations.ApiModelProperty;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class MemberListRequest {
-
-	@NotNull
-	@ApiModelProperty(value = "그룹PK", example = "1", required = true)
-	private Long channelId;
-
-}
